+++ SignalR/Hosting/IRequest.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Specialized;
+using System.Security.Principal;
 
 namespace SignalR.Hosting
 {
@@ -10,5 +11,6 @@ namespace SignalR.Hosting
         NameValueCollection Headers { get; }
         NameValueCollection Form { get; }
         IRequestCookieCollection Cookies { get; }
+        IPrincipal User { get; }
     }
 }