+++ samples/SignalR.Hosting.AspNet.Samples/Raw/Raw.cs
@@ -22,7 +22,7 @@ namespace SignalR.Samples.Raw
 
             string user = GetUser(connectionId);
 
-            return Groups.Add(connectionId, "foo").ContinueWith(_ => 
+            return Groups.Add(connectionId, "foo").ContinueWith(_ =>
                    Connection.Broadcast(DateTime.Now + ": " + user + " joined")).Unwrap();
         }
 
@@ -32,7 +32,7 @@ namespace SignalR.Samples.Raw
             return Connection.Broadcast(DateTime.Now + ": " + GetUser(connectionId) + " disconnected");
         }
 
-        protected override Task OnReceivedAsync(string connectionId, string data)
+        protected override Task OnReceivedAsync(IRequest request, string connectionId, string data)
         {
             var message = JsonConvert.DeserializeObject<Message>(data);
 
@@ -47,7 +47,7 @@ namespace SignalR.Samples.Raw
                     });
                     break;
                 case MessageType.Send:
-                    Send(new
+                    Connection.Send(connectionId, new
                     {
                         type = MessageType.Send,
                         from = GetUser(connectionId),
@@ -58,7 +58,7 @@ namespace SignalR.Samples.Raw
                     string name = message.Value;
                     _clients[connectionId] = name;
                     _users[name] = connectionId;
-                    Send(new
+                    Connection.Send(connectionId, new
                     {
                         type = MessageType.Join,
                         data = message.Value
@@ -91,7 +91,7 @@ namespace SignalR.Samples.Raw
                     break;
             }
 
-            return base.OnReceivedAsync(connectionId, data);
+            return base.OnReceivedAsync(request, connectionId, data);
         }
 
         private string GetUser(string connectionId)